language: add an empty template for the daml assistant. (#6102)
This adds an empty template for the daml assistant. This is very helpful
for writing katacoda's, because otherwise you have to delete contained
source files in the `skeleton` project everytime as a first step. The
README in the `daml` directory is mainly there to get the daml directory
in the template.

CHANGELOG_BEGIN
CHANGELOG_END