more granular test exclusions on api-test-tool
This PR adds a new flag, `--exclude-regex`, to the Ledger API Test Tool.
The new flag can be repeated; each instance provides one regex. Tests
are matched against the provided regexes by constructing an identifier
using the name of the test suite and the identifier of the test,
separated by a colon.

This PR also deprecates (but does not remove, nor alter the behaviour
of) the `--exclude` flag, as its use-cases are completely superseded by
`--exclude-regex`.

Note that regexes have to be given in Java format.

This fixes #5837, and, as far as I can tell, supersedes #3928.

CHANGELOG_BEGIN
- [Ledger API Test Tool] add a `--exclude-regex` flag that allows more
  fine-grained exclusion of tests. Deprecate `--exclude` in favour of
  the new flag.
CHANGELOG_END