remove ec from LedgerTestSuite
`LedgerTestSuite` itself does not do anything with the `ExecutionContext`;
it is only used to provide it as an "ambient" implicit throughout each
test definition. Instead, with this PR, the `ExecutionContext` is
explicitly passed down as an additional argument to the `runTestCase`
closure.

For this not to change the behaviour of the program, we need to be
confident that the `ExecutionContext` is still the same. Let's first
look at how the `ExecutionContext` used to reach a test case. For any
execution of the test tool, there is only one `LedgerTestSuiteRunner`
created [by the main method][0], on which [we call][1] the
`verifyRequirementsAndRun` method. This in turn [calls][2] the `run`
method after checking that the given configuration is valid, and
[that][3] is where the `ExecutionContext` gets created. It is declared
as implicit so it gets picked up by the [`LedgerSession`][4] constructor
[call][5]. This (single) `LedgerSession` then [gets passed][6] to the
[constructor][7] of each individual `LedgerTestSuite`, which [extracts
the `ExecutionContext` from the session][8] to make it available to each
individual test created in the enclosing scope of a subclass when
calling the [`test`][9] method.

With this PR, the `ExecutionContext` is no longer extracted from the
`LedgerSession` in the `LedgerTestSuite` constructor, but is instead
threaded through to the `apply` method of `LedgerTestCase` through the
`run` (different signature) [call][10] for the individual test, which
itself [calls][11] `start`, which then [calls][12] the `apply` method of
`LedgerTestCase`. Because each of these methods (`run` and `start` on
`LedgerTestSuiterunner` and `apply` on `LedgerTestCase`) already
expected an implicit `ExecutionContext`, no change is required on the
`LedgerTestSuiteRunner` side.

A handful of helper functions in various test suites also needed to be
changed to fetch the `ExecutionContext` implicitly from their calling
context rather than capturing it lexically from the enclosing
`LedgerTestSuite`. This is, again, the same `ExecutionContext` it used
to be, but instead of getting it from the `LedgerSession` they get it
from the `LedgerTestCase` closure.

This was originally part of #6314, but despite it being a pure
refactoring, it is quite a large (and somewhat subtle) changeset, so I
believe it is better to review/merge separately.

CHANGELOG_BEGIN
CHANGELOG_END

[0]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/LedgerApiTestTool.scala#L125-L136
[1]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/LedgerApiTestTool.scala#L138
[2]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerTestSuiteRunner.scala#L155
[3]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerTestSuiteRunner.scala#L119
[4]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerSession.scala#L13-L16
[5]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerTestSuiteRunner.scala#L122
[6]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerTestSuiteRunner.scala#L123
[7]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerTestSuite.scala#L13
[8]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerTestSuite.scala#L20
[9]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerTestSuite.scala#L22-L32
[10]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerTestSuiteRunner.scala#L134
[11]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerTestSuiteRunner.scala#L114
[12]: https://github.com/digital-asset/daml/blob/53bddb54d752153727ab990b50239bffb53d0ef6/ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/LedgerTestSuiteRunner.scala#L60