match test as prefix instead of suite
This is an alternative implementation of #6225. See discussions there
for more detail on what I felt was wrong with my first attempt.

I've decided to make a separate PR because this is different enough to
warrant separate reviews. All commits but the latest one are (intended
to be) pure refactoring; I strongly recommend reviewing one commit at a
time.

CHANGELOG_BEGIN
- [Ledger API Test Tool] ``--exclude`` and ``--include`` now match the
  full test name as a prefix, rather than just suite names. Test name is
  built by combining the suite name with a test identifier, so this
  change should be fully backwards compatible. Run with ``--list-all``
  to list all tests (as opposed to just the test suites with
  ``--list``).
CHANGELOG_END