reduce default concurrency of api-test-tool
This PR makes two small (?) changes to the Ledger API Test Tool. First,
it realizes the original intention that tests should run in a randomized
order, by explicily shuffling the list of tests (after selection through
include/exclude etc.). The existing implementation relies on the
ordering of `Map#values`, which, while not obviously connected to either
insertion order or sort order on the keys, is nonetheless deterministic
and hence results in all our tests always running in the same order (for
the same selection).

Second, because shuffling the tests results in a very flaky build, this
PR reduces the default concurrency from "as many threads as there are
CPU cores on this machine" to 1. On my machine this reduces the
performance of `//ledger/sandbox:conformance-test-static-time-h2database`
from running in about 30 seconds to running in about 55 seconds, so this
is definitely not negligible. However, trying to run with concurrency 2
on my machine yields a failure rate of about 20%, whereas setting
concurrency to 1 succeeded 40 times in a row.

I believe the underlying issue is that, after the test tool has been
implemented, the ledger API has been enhanced to support backpressure.
This is, in general, a good thing, but the test tool has apparently not
been updated to take it into account (say, with appropriate retry
strategies). Implementing that would be better than limiting
concurrency, but it would also be significantly more effort for me.

CHANGELOG_BEGIN
- [Ledger API Test Tool] Change default concurrency from "number of
  cores" to 1. You can still set explicit concurrency settings with
  ``--concurrent-test-runs INT``.
CHANGELOG_END