clarify date encoding in json-lf docs (#6388)
In #6378, I erroneously removed what I (and two other people who looked
at the document) thought was a typo. It turns out it wasn't, but I still
believe the existing form was a bit confusing. This is my attempt at
reintroducing and clarifying the original meaning.

CHANGELOG_BEGIN
CHANGELOG_END