Trigger service: Write packages to database if we have one
This is needed to recover state after the service shuts down or crashes.
We add a method to the RunningTriggerDao to persistPackages. This only
does something in the case of a DbTriggerDao. In any case the Server
keeps a package map in memory as it's required to construct a trigger runner.

I only write the main package dalf for now until I figure out how to fix my
traversal of the package list.

changelog_begin
changelog_end