remove temp linux nodes
This is the last step of the plan outlined in #6405. As of opening this
PR, "old" nodes are back up, "temp" nodes are disabled at the Azure
level, and there is no job running on either (:thinking:). In other
words, this can be deployed as soon as it gets a stamp.

CHANGELOG_BEGIN
CHANGELOG_END