Fix contains check in ConcurrentCompiledPackages (#6471)
* Fix contains check in ConcurrentCompiledPackages

`contains` searches for a _value_ not a _key_. We have a package id
here in both cases which is clearly a key.

I am not exactly clear on what exactly happens if you hit this bug. I
believe it’s just a performance issue so probably hard to write a test
for.

I did take a brief look at whether we can make this
typesafe (`contains` accept an `Object` which is how this typechecks)
and apparently calling `asScala` and then using
`scala.collection.concurrent.Map` should work but I don’t know if this
has performance implications or if there is another reason why we
didn’t do this. Happy to make the change if someone tells me this is
the right thing to do.

changelog_begin
changelog_end

* Switch to Scala’s concurrent map which doesn’t pretend types are bad

changelog_begin
changelog_end