bindings-rxjava: Return the `Disposable` from `Bot.wire`. (#6476)
* bindings-rxjava: Return the `Disposable` from `Bot.wire`.

And in `BotTest`, dispose of the connection.

We are seeing some spurious errors in runs of BotTest due to connections
shutting down in the wrong order. By explicitly disposing of the wiring
before shutting down the ledger client, we can hopefully suppress these
errors.

They're not causing test failures, but they do often obscure the real
failure.

CHANGELOG_BEGIN
- [RxJava Bindings] `Bot.wire` and `Bot.wireSimple` now return a
  `Disposable`, which can be used to shut down the flows. You are
  encouraged to call `.dispose()` before terminating the client.
CHANGELOG_END

* bindings-rxjava: Don't run `awaitTermination` in a loop.

When shutting down the ledger client, instead of calling
`channel.awaitTermination` in a loop, we just call it once with an
approximately-infinite timeout.

* bindings-rxjava: Use the loan pattern for disposing in BotTest.

* bindings-rxjava: Simplify BotTest a little further.